+++ src/Adapter/Server/_Server.mjs
@@ -1,3 +1,3 @@
 /**
- * @typedef {{listen_interface?: string, listen_https_port?: number, listen_http_port?: number, redirect_http_to_https?: boolean, redirect_http_to_https_port?: number, redirect_http_to_https_status_code?: number, https_certificate?: string, https_key?: string, https_dhparam?: string, no_date?: boolean, no_referrer?: boolean}} Server
+ * @typedef {{listen_interface?: string, listen_https_port?: number, listen_http_port?: number, redirect_http_to_https?: boolean, redirect_http_to_https_port?: number, redirect_http_to_https_status_code?: number, https_certificate?: string, https_key?: string, https_dhparam?: string, no_date?: boolean, no_referrer?: boolean, forwarded_headers?: boolean}} Server
  */