+++ src/Service/Server/Command/RunServerCommand.mjs
@@ -4,7 +4,7 @@ import { HEADER_REFERRER_POLICY } from "../../../Adapter/Header/HEADER.mjs";
 import { HttpServerResponse } from "../../../Adapter/Server/HttpServerResponse.mjs";
 import { PROTOCOL_HTTPS } from "../../../Adapter/Protocol/PROTOCOL.mjs";
 import { REFERRER_POLICY_NO_REFERRER } from "../../../Adapter/Referrer/REFERRER_POLICY.mjs";
-import { SERVER_DEFAULT_LISTEN_HTTP_PORT, SERVER_DEFAULT_LISTEN_HTTPS_PORT, SERVER_DEFAULT_NO_DATE, SERVER_DEFAULT_NO_REFERRER, SERVER_DEFAULT_REDIRECT_HTTP_TO_HTTPS, SERVER_DEFAULT_REDIRECT_HTTP_TO_HTTPS_PORT, SERVER_DEFAULT_REDIRECT_HTTP_TO_HTTPS_STATUS_CODE, SERVER_LISTEN_HTTP_PORT_DISABLED, SERVER_LISTEN_HTTPS_PORT_DISABLED } from "../../../Adapter/Server/SERVER.mjs";
+import { SERVER_DEFAULT_FORWARDED_HEADERS, SERVER_DEFAULT_LISTEN_HTTP_PORT, SERVER_DEFAULT_LISTEN_HTTPS_PORT, SERVER_DEFAULT_NO_DATE, SERVER_DEFAULT_NO_REFERRER, SERVER_DEFAULT_REDIRECT_HTTP_TO_HTTPS, SERVER_DEFAULT_REDIRECT_HTTP_TO_HTTPS_PORT, SERVER_DEFAULT_REDIRECT_HTTP_TO_HTTPS_STATUS_CODE, SERVER_LISTEN_HTTP_PORT_DISABLED, SERVER_LISTEN_HTTPS_PORT_DISABLED } from "../../../Adapter/Server/SERVER.mjs";
 import { STATUS_CODE_400, STATUS_CODE_404, STATUS_CODE_500 } from "../../../Adapter/Status/STATUS_CODE.mjs";
 
 /** @typedef {import("../../../Adapter/Server/handleRequest.mjs").handleRequest} handleRequest */
@@ -63,6 +63,7 @@ export class RunServerCommand {
         const https_dhparam = server?.https_dhparam ?? null;
         const no_date = server?.no_date ?? SERVER_DEFAULT_NO_DATE;
         const no_referrer = server?.no_referrer ?? SERVER_DEFAULT_NO_REFERRER;
+        const forwarded_headers = server?.forwarded_headers ?? SERVER_DEFAULT_FORWARDED_HEADERS;
 
         const https = listen_https_port !== SERVER_LISTEN_HTTPS_PORT_DISABLED && https_certificate !== null && https_key !== null;
         const http = listen_http_port !== SERVER_LISTEN_HTTP_PORT_DISABLED;
@@ -81,7 +82,8 @@ export class RunServerCommand {
                 redirect_http_to_https_port,
                 redirect_http_to_https_status_code,
                 no_date,
-                no_referrer
+                no_referrer,
+                forwarded_headers
             );
         };
 
@@ -154,9 +156,10 @@ export class RunServerCommand {
      * @param {number} redirect_http_to_https_status_code
      * @param {boolean} no_date
      * @param {boolean} no_referrer
+     * @param {boolean} forwarded_headers
      * @returns {Promise<void>}
      */
-    async #handleRequest(req, res, handle_request, redirect_http_to_https, redirect_http_to_https_port, redirect_http_to_https_status_code, no_date, no_referrer) {
+    async #handleRequest(req, res, handle_request, redirect_http_to_https, redirect_http_to_https_port, redirect_http_to_https_status_code, no_date, no_referrer, forwarded_headers) {
         res.sendDate = !no_date;
 
         if (no_referrer) {
@@ -167,7 +170,8 @@ export class RunServerCommand {
         try {
             request = await this.#server_service.mapRequest(
                 req,
-                res
+                res,
+                forwarded_headers
             );
         } catch (error) {
             console.error(error);